user, city and ticket services/controllers were added

+++ as_impl/src/main/java/ru/mano/aviasales/entity/City.java
@@ -7,10 +7,10 @@ import lombok.Data;
 @Data
 public class City extends AbstractEntityParent {
     private String name;
-    private int x, y;
+    private double x, y;
 
 
-    public City(int id, String name, int x, int y) {
+    public City(int id, String name, double x, double y) {
         super(id);
         this.name = name;
         this.x = x;
